+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - 废除 `f-alert` 组件 `overflow` 配置项
 - 新增 `f-alert` 组件 `alert-list` 配置项
 
+**修复**
+
+- 修复 `f-checkbox` 组件 `border` 状态样式问题
+
 ## 0.13.2 (2022-12-29)
 
 - 修复 `f-ripple` 组件失效问题